Use one single Flags struct.
