+++ src/lib.rs
@@ -14,13 +14,13 @@
 //!
 //! # Usage
 //! [`WordFilter`]s must be created at compile-time using the
-//! [`word_filter_codegen`](https://docs.rs/word_filter_codegen) crate. The generated code can then
+//! [`word_filter::codegen`](https://docs.rs/word_filter::codegen) crate. The generated code can then
 //! be [`include!`]ed and used.
 //!
 //! ## Example
 //! For example, a simple [`WordFilter`] can be generated by the following.
 //!
-//! First, add both the `word_filter` and `word_filter_codegen` crates to the `Cargo.toml`
+//! First, add both the `word_filter` and `word_filter::codegen` crates to the `Cargo.toml`
 //! `[dependencies]` and `[build-dependencies]` lists respectively.
 //!
 //! ``` toml
@@ -29,20 +29,20 @@
 //! word_filter = "0.7.0"
 //!
 //! [build-dependencies]
-//! word_filter_codegen = "0.7.0"
+//! word_filter::codegen = "0.7.0"
 //! ...
 //! ```
 //!
 //! Next, generate the [`WordFilter`] in the `build.rs` file.
 //!
-//! ``` ignore
+//! ```
 //! use std::{
 //!     env,
 //!     fs::File,
 //!     io::{BufWriter, Write},
 //!     path::Path,
 //! };
-//! use word_filter_codegen::{Visibility, WordFilterGenerator};
+//! use word_filter::codegen::{Visibility, WordFilterGenerator};
 //!
 //! fn main() {
 //!     let path = Path::new(&env::var("OUT_DIR").unwrap()).join("codegen.rs");
@@ -99,7 +99,7 @@ use unicode_segmentation::UnicodeSegmentation;
 /// These are of the form `(<source string>, <alias string>)`.
 ///
 /// `WordFilter`s are constructed at compile-time using the
-/// [`word_filter_codegen`](https://docs.rs/word_filter_codegen) crate. See crate-level
+/// [`word_filter::codegen`](https://docs.rs/word_filter::codegen) crate. See crate-level
 /// documentation for further information.
 #[derive(Debug)]
 pub struct WordFilter<'a, const N: usize> {
@@ -184,7 +184,7 @@ impl<'a, const N: usize> WordFilter<'a, N> {
     ///     io::{BufWriter, Write},
     ///     path::Path,
     /// };
-    /// use word_filter_codegen::{Visibility, WordFilterGenerator};
+    /// use word_filter::codegen::{Visibility, WordFilterGenerator};
     ///
     /// fn main() {
     ///     let path = Path::new(&env::var("OUT_DIR").unwrap()).join("codegen.rs");
@@ -232,7 +232,7 @@ impl<'a, const N: usize> WordFilter<'a, N> {
     ///     io::{BufWriter, Write},
     ///     path::Path,
     /// };
-    /// use word_filter_codegen::{Visibility, WordFilterGenerator};
+    /// use word_filter::codegen::{Visibility, WordFilterGenerator};
     ///
     /// fn main() {
     ///     let path = Path::new(&env::var("OUT_DIR").unwrap()).join("codegen.rs");
@@ -276,7 +276,7 @@ impl<'a, const N: usize> WordFilter<'a, N> {
     ///     io::{BufWriter, Write},
     ///     path::Path,
     /// };
-    /// use word_filter_codegen::{Visibility, WordFilterGenerator};
+    /// use word_filter::codegen::{Visibility, WordFilterGenerator};
     ///
     /// fn main() {
     ///     let path = Path::new(&env::var("OUT_DIR").unwrap()).join("codegen.rs");
@@ -351,7 +351,7 @@ impl<'a, const N: usize> WordFilter<'a, N> {
     ///     io::{BufWriter, Write},
     ///     path::Path,
     /// };
-    /// use word_filter_codegen::{Visibility, WordFilterGenerator};
+    /// use word_filter::codegen::{Visibility, WordFilterGenerator};
     ///
     /// fn main() {
     ///     let path = Path::new(&env::var("OUT_DIR").unwrap()).join("codegen.rs");
